+++ iad/model/client.py
@@ -1,13 +1,21 @@
 from datetime import datetime
 from io import BytesIO
-from aiohttp import ClientSession
 from os import path
-from ..errors import JsonDecodeException, RequestError, InvalidContentType
+from typing import Union
+
+from aiohttp import ClientSession
+
+from ..errors import InvalidContentType, JsonDecodeException, RequestError
 from ..ratelimit import RateLimiter
 from ..request import Request
 from .upload import Upload
 
+
 class ImADev:
+    """
+    Main class for iad.py
+    """
+
     DISABLED = ('php', 'html', 'js', 'css', 'ts')
     def __init__(self, token, session = None):
         assert isinstance(token, str), f"token argument expected str, got {token.__class__.__name__}"
@@ -16,9 +24,17 @@ class ImADev:
         self._http = RateLimiter(session or ClientSession())
 
     def __repr__(self):
-        return f"<ImADev token={self.__token}>"
+        return f"<ImADev token={self.__token} http={self._http}>"
 
-    async def upload(self, data, file_format: str = None):
+    async def upload(self, data, file_format: str = None) -> Upload:
+        """
+        Upload a file to im-a-dev.xyz
+
+        :param data: either :class:`io.BytesIO` file buffer, :class:`bytes` file bytes, or `:class:`str` file path
+        :param file_format: must be :class:`str`, can be ``None`` if data is str.
+        :return: Returns a :class:`iad.model.Upload`
+        """
+
         if isinstance(data, str):
             assert isinstance(data, str), "file_format argument not specified"
             assert path.isfile(data), f"File not found: {data}"
@@ -51,7 +67,33 @@ class ImADev:
         except:
             raise RequestError(json)
 
-    async def get_upload(self, filename: str):
+
+    async def delete(self, filename: Union[Upload, str]) -> bool:
+        """
+        :param filename: must be either a :class:`iad.model.Upload` or a :class:`str` representing the file/filename
+        :return: Returns a :class:`bool` if it was succsesful or not.
+        """
+
+        assert type(filename) in [Upload, str], f"filename expected str or upload, got {filename.__class__.__name__}"
+
+        req = Request(self._http, 'upload.php', 'POST', data={
+            'token': self.__token,
+            'endpoint': 'delete',
+            'filename': filename
+        })
+
+        try:
+            json = await req.json()
+        except:
+            raise JsonDecodeException(await req.read())
+
+        return (json['http_code'] == 200)
+
+    async def get_upload(self, filename: str) -> Upload:
+        """
+        :param filename: Name of the file to get
+        :return: Returns a :class:`iad.model.Upload` of the fetched upload.
+        """
         if not isinstance(filename, str):
             raise TypeError(f'filename argument expected str, got {filename.__class__.__name__}')